Change parse_module to return Vec<VimNode>, get rid of VimModule type
