+++ lib/src/data.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, PartialEq)]
+pub enum VimNode {
+    StandaloneDocComment(String),
+    Function { name: String, doc: Option<String> },
+}
